+++ vexcl/gather.hpp
@@ -31,6 +31,10 @@ THE SOFTWARE.
  * \brief  Gather scattered points from OpenCL device vector.
  */
 
+#include <vector>
+#include <numeric>
+#include <cassert>
+
 #include <vexcl/operations.hpp>
 #include <vexcl/vector.hpp>
 
@@ -52,14 +56,14 @@ class gather {
             column_owner owner(part);
 
             for(auto i = indices.begin(); i != indices.end(); ++i) {
-                uint d = owner(*i);
+                size_t d = owner(*i);
                 *i -= part[d];
                 ++ptr[d + 1];
             }
 
             std::partial_sum(ptr.begin(), ptr.end(), ptr.begin());
 
-            for(uint d = 0; d < queue.size(); d++) {
+            for(unsigned d = 0; d < queue.size(); d++) {
                 cl::Context context = qctx(queue[d]);
 
                 if (size_t n = ptr[d + 1] - ptr[d]) {
@@ -71,7 +75,7 @@ class gather {
                 }
             }
 
-            for(uint d = 0; d < queue.size(); d++)
+            for(unsigned d = 0; d < queue.size(); d++)
                 if (ptr[d + 1] - ptr[d]) ev[d].wait();
         }
 
@@ -81,7 +85,7 @@ class gather {
 
             static kernel_cache cache;
 
-            for(uint d = 0; d < queue.size(); d++) {
+            for(unsigned d = 0; d < queue.size(); d++) {
                 cl::Context context = qctx(queue[d]);
                 cl::Device  device  = qdev(queue[d]);
 
@@ -116,7 +120,7 @@ class gather {
                     size_t w_size = kernel->second.wgsize;
                     size_t g_size = alignup(n, w_size);
 
-                    uint pos = 0;
+                    unsigned pos = 0;
                     kernel->second.kernel.setArg(pos++, n);
                     kernel->second.kernel.setArg(pos++, src(d));
                     kernel->second.kernel.setArg(pos++, idx[d]);
@@ -131,7 +135,7 @@ class gather {
                 }
             }
 
-            for(uint d = 0; d < queue.size(); d++)
+            for(unsigned d = 0; d < queue.size(); d++)
                 if (ptr[d + 1] - ptr[d]) ev[d].wait();
         }
     private:
